requirements.txt
annotated-types==0.7.0
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
QuantLib==1.39
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
